Updated Functionalities

+++ models/linkTree.js
@@ -4,46 +4,34 @@ const connection = require('../config/database');
 const createLinkTree = (linkTree, callback) => {
   const query = 'INSERT INTO LinkTree (url, tree, userAccountId) VALUES (?, ?, ?)';
   connection.query(query, [linkTree.url, JSON.stringify(linkTree.tree), linkTree.userAccountId], (err, results) => {
-    if (err) {
-      console.error('Error creating link tree:', err);
-      return callback(err);
-    }
-    callback(null, { id: results.insertId });
+    if (err) return callback(err);
+    callback(null, results);
   });
 };
 
-// Find a link tree by URL
-const findLinkTreeByUrl = (url, callback) => {
-  const query = 'SELECT * FROM LinkTree WHERE url = ?';
-  connection.query(query, [url], (err, results) => {
-    if (err) {
-      console.error('Error finding link tree by URL:', err);
-      return callback(err);
-    }
+// Find a link tree by URL and user ID
+const findLinkTreeByUrl = (url, userId, callback) => {
+  const query = 'SELECT * FROM LinkTree WHERE url = ? AND userAccountId = ?';
+  connection.query(query, [url, userId], (err, results) => {
+    if (err) return callback(err);
     callback(null, results[0]);
   });
 };
 
 // Update a link tree
 const updateLinkTree = (linkTree, callback) => {
-  const query = 'UPDATE LinkTree SET url = ?, tree = ? WHERE id = ?';
-  connection.query(query, [linkTree.url, JSON.stringify(linkTree.tree), linkTree.id], (err, results) => {
-    if (err) {
-      console.error('Error updating link tree:', err);
-      return callback(err);
-    }
+  const query = 'UPDATE LinkTree SET tree = ? WHERE id = ? AND userAccountId = ?';
+  connection.query(query, [JSON.stringify(linkTree.tree), linkTree.id, linkTree.userAccountId], (err, results) => {
+    if (err) return callback(err);
     callback(null, results);
   });
 };
 
 // Delete a link tree
-const deleteLinkTree = (id, callback) => {
-  const query = 'DELETE FROM LinkTree WHERE id = ?';
-  connection.query(query, [id], (err, results) => {
-    if (err) {
-      console.error('Error deleting link tree:', err);
-      return callback(err);
-    }
+const deleteLinkTree = (id, userId, callback) => {
+  const query = 'DELETE FROM LinkTree WHERE id = ? AND userAccountId = ?';
+  connection.query(query, [id, userId], (err, results) => {
+    if (err) return callback(err);
     callback(null, results);
   });
 };
